requirements.txt
Flask==2.1.3
multitimer==0.3
numpy
python-dotenv
opencv-python-headless
tensorflow-cpu==2.9.1
